+++ src/main/java/dao/Dao.java
@@ -4,18 +4,14 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.ArrayList;
 
-/**
- * createAt Dec 24, 2020
- *
- * @author Đỗ Tuấn Anh <daclip26@gmail.com>
- */
+
 public abstract class Dao<T> {
 
     Connection conn = Database.getInstance().getConnection();
 
     public abstract ArrayList<T> getAll() throws SQLException;
 
-    public abstract T get(int id) throws SQLException;
+    public abstract T get(Long id) throws SQLException;
 
     public abstract void save(T t) throws SQLException;
 
@@ -23,5 +19,5 @@ public abstract class Dao<T> {
 
     public abstract void delete(T t) throws SQLException;
 
-    public abstract void deleteById(int id) throws SQLException;
+    public abstract void deleteById(Long id) throws SQLException;
 }
